move helpers to event, move stuff to env, general cleanup
