+++ .env.example
@@ -1,4 +1,4 @@
-DEBUG_FLAG_ENABLED=1
+DEBUG_FLAG_ENABLED=0
 
 GH_DISCUSSION_CATEGORY_ID="DIC_kwDOE7ysSc4CAC0o"
 GH_ACCESS_TOKEN="💖"
@@ -7,3 +7,6 @@ GH_ACCESS_TOKEN="💖"
 # It is accessed at runtime via svelte.config.js preprocessing.
 # If a value is not provided, analytics are ignored.
 VERCEL_ANALYTICS_ID=""
+ 
+GCAL_API_KEY=""
+GCAL_ID=""
